requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
phonenumbers==8.13.23
email-validator==2.0.0
python-dotenv==1.0.0
Pillow==10.0.0
Werkzeug==2.2.3
SQLAlchemy==2.0.23
WTForms==3.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
